add in the full clone getter
