+++ beacon-chain/state/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["types.go"],
+    srcs = [
+        "getters.go",
+        "setters.go",
+        "types.go",
+    ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
@@ -10,5 +14,7 @@ go_library(
         "//shared/bytesutil:go_default_library",
         "//shared/hashutil:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
+        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
+        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
     ],
 )
