+++ beacon-chain/state/setters.go
@@ -1,0 +1,1 @@
+package state
